move pod operation functions to pod_utils.go
