+++ test/integration/dataplane_test.go
@@ -22,7 +22,7 @@ import (
 	operatorv1alpha1 "github.com/kong/gateway-operator/apis/v1alpha1"
 	"github.com/kong/gateway-operator/controllers"
 	"github.com/kong/gateway-operator/internal/consts"
-	k8sresources "github.com/kong/gateway-operator/internal/utils/kubernetes/resources"
+	k8sutils "github.com/kong/gateway-operator/internal/utils/kubernetes"
 	testutils "github.com/kong/gateway-operator/internal/utils/test"
 )
 
@@ -86,7 +86,7 @@ func TestDataplaneEssentials(t *testing.T) {
 	require.Len(t, deployments, 1, "There must be only one ControlPlane deployment")
 	deployment := &deployments[0]
 
-	controllerContainer := k8sresources.GetPodContainerByName(
+	controllerContainer := k8sutils.GetPodContainerByName(
 		&deployment.Spec.Template.Spec, consts.DataPlaneProxyContainerName)
 	require.NotNil(t, controllerContainer)
 	envs := controllerContainer.Env
@@ -229,7 +229,7 @@ func TestDataPlaneUpdate(t *testing.T) {
 	deployment := &deployments[0]
 
 	t.Logf("verifying environment variable TEST_ENV in deployment before update")
-	container := k8sresources.GetPodContainerByName(&deployment.Spec.Template.Spec, consts.DataPlaneProxyContainerName)
+	container := k8sutils.GetPodContainerByName(&deployment.Spec.Template.Spec, consts.DataPlaneProxyContainerName)
 	require.NotNil(t, container)
 	testEnv := getEnvValueByName(container.Env, "TEST_ENV")
 	require.Equal(t, "before_update", testEnv)
@@ -251,7 +251,7 @@ func TestDataPlaneUpdate(t *testing.T) {
 		require.Len(t, deployments, 1, "There must be only one ControlPlane deployment")
 		deployment := &deployments[0]
 
-		container := k8sresources.GetPodContainerByName(&deployment.Spec.Template.Spec, consts.DataPlaneProxyContainerName)
+		container := k8sutils.GetPodContainerByName(&deployment.Spec.Template.Spec, consts.DataPlaneProxyContainerName)
 		require.NotNil(t, container)
 		testEnv := getEnvValueByName(container.Env, "TEST_ENV")
 		t.Logf("Tenvironment variable TEST_ENV is now %s in deployment", testEnv)
